Improve the accuracy of the toolchain file
The lib paths are populated according to the LIB variable from the
developer console, and the lib paths and include paths are now in the
order they appear in the developer console.

The fms-compatibility-version flag is now added to the compile flags to
have the correct defines for system headers. This requires the version
of cl.exe to be passed, but there is no easy way to find that out from
the virtual environment description provided by GitHub. You must run an
action that prints that info to find it out.